requirements.txt
python-frontmatter==1.0.0
pyyaml==6.0.1
